requirements.txt
openai
PyPDF2
requests
pandas
openpyxl
python-dotenv
pdfplumber
pymupdf
pix2tex
numpy
pillow
torch
transformers